requirements.txt
# Django
Django==3.2.0
djangorestframework
django-cors-headers
django-crispy-forms
django-extensions
django-filter
django-pagedown
django-rest-swagger
djangorestframework-jwt
django-environ
django-jsonfield
dj-static
appnope
backcall
certifi
chardet
coreapi
coreschema
decorator
djoser
idna
ipdb
ipython
ipython-genutils
itypes
jedi
Jinja2
parso
pexpect
pickleshare
Pillow
prompt-toolkit
psycopg2-binary
ptyprocess
Pygments
PyJWT
pypng
pytz
requests
simplejson
six
sqlparse
traitlets
uritemplate
urllib3
wcwidth
gunicorn
asgiref
pycodestyle
autopep8
Unipath
dj-database-url
python-decouple
whitenoise
more-itertools
django-simple-history

wget
xlwt

# Redis
django-redis
redis
django-cacheops

# Mysql
django-mysql
mysqlclient
